check for too long LEB128 sequences with tests
Testing this thoroughly ended up needing unit tests for LEB128
encoding and decoding as a standalone unit since otherwise you
need to work with extremely long sequences of zeros. These could
happen in principle, but this lets us test our LEB128 functions
without that. The leb128{e,d} programs could be useful on their
own, but it's hard to think of a circumstance.